[Tui] Deduplicate terminal escape sequences into a trait
